move XtextResourceLocator
