+++ plugins/org.eclipse.n4js/src/org/eclipse/n4js/scoping/builtin/ConfiguredResourceSetProvider.java
@@ -14,8 +14,8 @@ import java.lang.reflect.InvocationTargetException;
 
 import org.eclipse.emf.ecore.resource.ResourceSet;
 import org.eclipse.emf.ecore.resource.impl.ResourceSetImpl.ResourceLocator;
+import org.eclipse.n4js.resource.XtextResourceLocator;
 import org.eclipse.n4js.xtext.resourceset.EmptyAuthorityAddingNormalizer;
-import org.eclipse.n4js.xtext.resourceset.XtextResourceLocator;
 import org.eclipse.xtext.resource.SynchronizedXtextResourceSet;
 import org.eclipse.xtext.util.UriExtensions;
